requirements.txt
# Core Dependencies
PyQt6>=6.4.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-image>=0.21.0
imagehash>=4.3.0
rawpy>=0.18.0

# Build Tools
cx-Freeze>=6.15.0

# Future Features (Optional)
#face-recognition>=1.3.0  # For face detection
# geopy>=2.3.0  # For geotag processing
# pytesseract>=0.3.10  # For OCR in document scanning
